Split query related functions out into a separate file
