+++ src/core/providers/arcgis/qgsarcgisrestutils.cpp
@@ -16,14 +16,10 @@
 #include "qgsarcgisrestutils.h"
 #include "qgsfields.h"
 #include "qgslogger.h"
-#include "qgsnetworkaccessmanager.h"
 #include "qgsrectangle.h"
-#include "qgsfeedback.h"
 #include "qgspallabeling.h"
 #include "qgssymbol.h"
 #include "qgssymbollayer.h"
-#include "qgsauthmanager.h"
-#include "qgssettings.h"
 #include "qgslinesymbollayer.h"
 #include "qgsfillsymbollayer.h"
 #include "qgsrenderer.h"
@@ -31,8 +27,6 @@
 #include "qgssinglesymbolrenderer.h"
 #include "qgscategorizedsymbolrenderer.h"
 #include "qgsvectorlayerlabeling.h"
-#include "qgsapplication.h"
-#include "qgsmessagelog.h"
 #include "qgscircularstring.h"
 #include "qgsmulticurve.h"
 #include "qgspolygon.h"
@@ -41,13 +35,6 @@
 #include "qgsgeometryengine.h"
 #include "qgsmultisurface.h"
 #include "qgsmultipoint.h"
-
-#include <QEventLoop>
-#include <QNetworkRequest>
-#include "qgsblockingnetworkrequest.h"
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QImageReader>
 
 QVariant::Type QgsArcGisRestUtils::convertFieldType( const QString &esriFieldType )
 {
@@ -398,184 +385,6 @@ QgsCoordinateReferenceSystem QgsArcGisRestUtils::convertSpatialReference( const 
   }
   return crs;
 }
-
-///@cond PRIVATE
-QVariantMap QgsArcGisRestUtils::getServiceInfo( const QString &baseurl, const QString &authcfg, QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders )
-{
-  // http://sampleserver5.arcgisonline.com/arcgis/rest/services/Energy/Geology/FeatureServer?f=json
-  QUrl queryUrl( baseurl );
-  QUrlQuery query( queryUrl );
-  query.addQueryItem( QStringLiteral( "f" ), QStringLiteral( "json" ) );
-  queryUrl.setQuery( query );
-  return queryServiceJSON( queryUrl, authcfg, errorTitle, errorText, requestHeaders );
-}
-
-QVariantMap QgsArcGisRestUtils::getLayerInfo( const QString &layerurl, const QString &authcfg, QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders )
-{
-  // http://sampleserver5.arcgisonline.com/arcgis/rest/services/Energy/Geology/FeatureServer/1?f=json
-  QUrl queryUrl( layerurl );
-  QUrlQuery query( queryUrl );
-  query.addQueryItem( QStringLiteral( "f" ), QStringLiteral( "json" ) );
-  queryUrl.setQuery( query );
-  return queryServiceJSON( queryUrl, authcfg, errorTitle, errorText, requestHeaders );
-}
-
-QVariantMap QgsArcGisRestUtils::getObjectIds( const QString &layerurl, const QString &authcfg, QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders, const QgsRectangle &bbox )
-{
-  // http://sampleserver5.arcgisonline.com/arcgis/rest/services/Energy/Geology/FeatureServer/1/query?where=1%3D1&returnIdsOnly=true&f=json
-  QUrl queryUrl( layerurl + "/query" );
-  QUrlQuery query( queryUrl );
-  query.addQueryItem( QStringLiteral( "f" ), QStringLiteral( "json" ) );
-  query.addQueryItem( QStringLiteral( "where" ), QStringLiteral( "1=1" ) );
-  query.addQueryItem( QStringLiteral( "returnIdsOnly" ), QStringLiteral( "true" ) );
-  if ( !bbox.isNull() )
-  {
-    query.addQueryItem( QStringLiteral( "geometry" ), QStringLiteral( "%1,%2,%3,%4" )
-                        .arg( bbox.xMinimum(), 0, 'f', -1 ).arg( bbox.yMinimum(), 0, 'f', -1 )
-                        .arg( bbox.xMaximum(), 0, 'f', -1 ).arg( bbox.yMaximum(), 0, 'f', -1 ) );
-    query.addQueryItem( QStringLiteral( "geometryType" ), QStringLiteral( "esriGeometryEnvelope" ) );
-    query.addQueryItem( QStringLiteral( "spatialRel" ), QStringLiteral( "esriSpatialRelEnvelopeIntersects" ) );
-  }
-  queryUrl.setQuery( query );
-  return queryServiceJSON( queryUrl, authcfg, errorTitle, errorText, requestHeaders );
-}
-
-QVariantMap QgsArcGisRestUtils::getObjects( const QString &layerurl, const QString &authcfg, const QList<quint32> &objectIds, const QString &crs,
-    bool fetchGeometry, const QStringList &fetchAttributes,
-    bool fetchM, bool fetchZ,
-    const QgsRectangle &filterRect,
-    QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders, QgsFeedback *feedback )
-{
-  QStringList ids;
-  for ( int id : objectIds )
-  {
-    ids.append( QString::number( id ) );
-  }
-  QUrl queryUrl( layerurl + "/query" );
-  QUrlQuery query( queryUrl );
-  query.addQueryItem( QStringLiteral( "f" ), QStringLiteral( "json" ) );
-  query.addQueryItem( QStringLiteral( "objectIds" ), ids.join( QLatin1Char( ',' ) ) );
-  QString wkid = crs.indexOf( QLatin1Char( ':' ) ) >= 0 ? crs.split( ':' )[1] : QString();
-  query.addQueryItem( QStringLiteral( "inSR" ), wkid );
-  query.addQueryItem( QStringLiteral( "outSR" ), wkid );
-
-  query.addQueryItem( QStringLiteral( "returnGeometry" ), fetchGeometry ? QStringLiteral( "true" ) : QStringLiteral( "false" ) );
-
-  QString outFields;
-  if ( fetchAttributes.isEmpty() )
-    outFields = QStringLiteral( "*" );
-  else
-    outFields = fetchAttributes.join( ',' );
-  query.addQueryItem( QStringLiteral( "outFields" ), outFields );
-
-  query.addQueryItem( QStringLiteral( "returnM" ), fetchM ? QStringLiteral( "true" ) : QStringLiteral( "false" ) );
-  query.addQueryItem( QStringLiteral( "returnZ" ), fetchZ ? QStringLiteral( "true" ) : QStringLiteral( "false" ) );
-  if ( !filterRect.isNull() )
-  {
-    query.addQueryItem( QStringLiteral( "geometry" ), QStringLiteral( "%1,%2,%3,%4" )
-                        .arg( filterRect.xMinimum(), 0, 'f', -1 ).arg( filterRect.yMinimum(), 0, 'f', -1 )
-                        .arg( filterRect.xMaximum(), 0, 'f', -1 ).arg( filterRect.yMaximum(), 0, 'f', -1 ) );
-    query.addQueryItem( QStringLiteral( "geometryType" ), QStringLiteral( "esriGeometryEnvelope" ) );
-    query.addQueryItem( QStringLiteral( "spatialRel" ), QStringLiteral( "esriSpatialRelEnvelopeIntersects" ) );
-  }
-  queryUrl.setQuery( query );
-  return queryServiceJSON( queryUrl,  authcfg, errorTitle, errorText, requestHeaders, feedback );
-}
-
-QList<quint32> QgsArcGisRestUtils::getObjectIdsByExtent( const QString &layerurl, const QgsRectangle &filterRect, QString &errorTitle, QString &errorText, const QString &authcfg, const QgsStringMap &requestHeaders, QgsFeedback *feedback )
-{
-  QUrl queryUrl( layerurl + "/query" );
-  QUrlQuery query( queryUrl );
-  query.addQueryItem( QStringLiteral( "f" ), QStringLiteral( "json" ) );
-  query.addQueryItem( QStringLiteral( "where" ), QStringLiteral( "1=1" ) );
-  query.addQueryItem( QStringLiteral( "returnIdsOnly" ), QStringLiteral( "true" ) );
-  query.addQueryItem( QStringLiteral( "geometry" ), QStringLiteral( "%1,%2,%3,%4" )
-                      .arg( filterRect.xMinimum(), 0, 'f', -1 ).arg( filterRect.yMinimum(), 0, 'f', -1 )
-                      .arg( filterRect.xMaximum(), 0, 'f', -1 ).arg( filterRect.yMaximum(), 0, 'f', -1 ) );
-  query.addQueryItem( QStringLiteral( "geometryType" ), QStringLiteral( "esriGeometryEnvelope" ) );
-  query.addQueryItem( QStringLiteral( "spatialRel" ), QStringLiteral( "esriSpatialRelEnvelopeIntersects" ) );
-  queryUrl.setQuery( query );
-  const QVariantMap objectIdData = queryServiceJSON( queryUrl, authcfg, errorTitle, errorText, requestHeaders, feedback );
-
-  if ( objectIdData.isEmpty() )
-  {
-    return QList<quint32>();
-  }
-
-  QList<quint32> ids;
-  const QVariantList objectIdsList = objectIdData[QStringLiteral( "objectIds" )].toList();
-  ids.reserve( objectIdsList.size() );
-  for ( const QVariant &objectId : objectIdsList )
-  {
-    ids << objectId.toInt();
-  }
-  return ids;
-}
-
-QByteArray QgsArcGisRestUtils::queryService( const QUrl &u, const QString &authcfg, QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders, QgsFeedback *feedback, QString *contentType )
-{
-  QUrl url = parseUrl( u );
-
-  QNetworkRequest request( url );
-  QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsArcGisRestUtils" ) );
-  for ( auto it = requestHeaders.constBegin(); it != requestHeaders.constEnd(); ++it )
-  {
-    request.setRawHeader( it.key().toUtf8(), it.value().toUtf8() );
-  }
-
-  QgsBlockingNetworkRequest networkRequest;
-  networkRequest.setAuthCfg( authcfg );
-  const QgsBlockingNetworkRequest::ErrorCode error = networkRequest.get( request, false, feedback );
-
-  if ( feedback && feedback->isCanceled() )
-    return QByteArray();
-
-  // Handle network errors
-  if ( error != QgsBlockingNetworkRequest::NoError )
-  {
-    QgsDebugMsg( QStringLiteral( "Network error: %1" ).arg( networkRequest.errorMessage() ) );
-    errorTitle = QStringLiteral( "Network error" );
-    errorText = networkRequest.errorMessage();
-    return QByteArray();
-  }
-
-  const QgsNetworkReplyContent content = networkRequest.reply();
-  if ( contentType )
-    *contentType = content.rawHeader( "Content-Type" );
-  return content.content();
-}
-
-QVariantMap QgsArcGisRestUtils::queryServiceJSON( const QUrl &url, const QString &authcfg, QString &errorTitle, QString &errorText, const QgsStringMap &requestHeaders, QgsFeedback *feedback )
-{
-  QByteArray reply = queryService( url, authcfg, errorTitle, errorText, requestHeaders, feedback );
-  if ( !errorTitle.isEmpty() )
-  {
-    return QVariantMap();
-  }
-  if ( feedback && feedback->isCanceled() )
-    return QVariantMap();
-
-  // Parse data
-  QJsonParseError err;
-  QJsonDocument doc = QJsonDocument::fromJson( reply, &err );
-  if ( doc.isNull() )
-  {
-    errorTitle = QStringLiteral( "Parsing error" );
-    errorText = err.errorString();
-    QgsDebugMsg( QStringLiteral( "Parsing error: %1" ).arg( err.errorString() ) );
-    return QVariantMap();
-  }
-  const QVariantMap res = doc.object().toVariantMap();
-  if ( res.contains( QStringLiteral( "error" ) ) )
-  {
-    const QVariantMap error = res.value( QStringLiteral( "error" ) ).toMap();
-    errorText = error.value( QStringLiteral( "message" ) ).toString();
-    errorTitle = QObject::tr( "Error %1" ).arg( error.value( QStringLiteral( "code" ) ).toString() );
-    return QVariantMap();
-  }
-  return res;
-}
-///@endcond
 
 QgsSymbol *QgsArcGisRestUtils::convertSymbol( const QVariantMap &symbolData )
 {
@@ -1089,335 +898,3 @@ QDateTime QgsArcGisRestUtils::convertDateTime( const QVariant &value )
   else
     return dt;
 }
-
-QUrl QgsArcGisRestUtils::parseUrl( const QUrl &url )
-{
-  QUrl modifiedUrl( url );
-  if ( modifiedUrl.toString().contains( QLatin1String( "fake_qgis_http_endpoint" ) ) )
-  {
-    // Just for testing with local files instead of http:// resources
-    QString modifiedUrlString = modifiedUrl.toString();
-    // Qt5 does URL encoding from some reason (of the FILTER parameter for example)
-    modifiedUrlString = QUrl::fromPercentEncoding( modifiedUrlString.toUtf8() );
-    modifiedUrlString.replace( QLatin1String( "fake_qgis_http_endpoint/" ), QLatin1String( "fake_qgis_http_endpoint_" ) );
-    QgsDebugMsg( QStringLiteral( "Get %1" ).arg( modifiedUrlString ) );
-    modifiedUrlString = modifiedUrlString.mid( QStringLiteral( "http://" ).size() );
-    QString args = modifiedUrlString.mid( modifiedUrlString.indexOf( '?' ) );
-    if ( modifiedUrlString.size() > 150 )
-    {
-      args = QCryptographicHash::hash( args.toUtf8(), QCryptographicHash::Md5 ).toHex();
-    }
-    else
-    {
-      args.replace( QLatin1String( "?" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "&" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "<" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( ">" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "'" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "\"" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( " " ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( ":" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "/" ), QLatin1String( "_" ) );
-      args.replace( QLatin1String( "\n" ), QLatin1String( "_" ) );
-    }
-#ifdef Q_OS_WIN
-    // Passing "urls" like "http://c:/path" to QUrl 'eats' the : after c,
-    // so we must restore it
-    if ( modifiedUrlString[1] == '/' )
-    {
-      modifiedUrlString = modifiedUrlString[0] + ":/" + modifiedUrlString.mid( 2 );
-    }
-#endif
-    modifiedUrlString = modifiedUrlString.mid( 0, modifiedUrlString.indexOf( '?' ) ) + args;
-    QgsDebugMsg( QStringLiteral( "Get %1 (after laundering)" ).arg( modifiedUrlString ) );
-    modifiedUrl = QUrl::fromLocalFile( modifiedUrlString );
-  }
-
-  return modifiedUrl;
-}
-
-
-void QgsArcGisRestUtils::adjustBaseUrl( QString &baseUrl, const QString name )
-{
-  QStringList parts = name.split( '/' );
-  QString checkString;
-  for ( const QString &part : parts )
-  {
-    if ( !checkString.isEmpty() )
-      checkString += QString( '/' );
-
-    checkString += part;
-    if ( baseUrl.indexOf( QRegularExpression( checkString.replace( '/', QLatin1String( "\\/" ) ) + QStringLiteral( "\\/?$" ) ) ) > -1 )
-    {
-      baseUrl = baseUrl.left( baseUrl.length() - checkString.length() - 1 );
-      break;
-    }
-  }
-}
-
-///@cond PRIVATE
-void QgsArcGisRestUtils::visitFolderItems( const std::function< void( const QString &, const QString & ) > &visitor, const QVariantMap &serviceData, const QString &baseUrl )
-{
-  QString base( baseUrl );
-  bool baseChecked = false;
-  if ( !base.endsWith( '/' ) )
-    base += QLatin1Char( '/' );
-
-  const QStringList folderList = serviceData.value( QStringLiteral( "folders" ) ).toStringList();
-  for ( const QString &folder : folderList )
-  {
-    if ( !baseChecked )
-    {
-      adjustBaseUrl( base, folder );
-      baseChecked = true;
-    }
-    visitor( folder, base + folder );
-  }
-}
-
-void QgsArcGisRestUtils::visitServiceItems( const std::function< void( const QString &, const QString & ) > &visitor, const QVariantMap &serviceData, const QString &baseUrl, const ServiceTypeFilter filter )
-{
-  QString base( baseUrl );
-  bool baseChecked = false;
-  if ( !base.endsWith( '/' ) )
-    base += QLatin1Char( '/' );
-
-  const QVariantList serviceList = serviceData.value( QStringLiteral( "services" ) ).toList();
-  for ( const QVariant &service : serviceList )
-  {
-    const QVariantMap serviceMap = service.toMap();
-    const QString serviceType = serviceMap.value( QStringLiteral( "type" ) ).toString();
-    if ( serviceType != QLatin1String( "MapServer" ) && serviceType != QLatin1String( "ImageServer" ) && serviceType != QLatin1String( "FeatureServer" ) )
-      continue;
-
-    // If the requested service type is raster, do not show vector-only services
-    if ( serviceType == QLatin1String( "FeatureServer" ) && filter == QgsArcGisRestUtils::Raster )
-      continue;
-
-    const QString serviceName = serviceMap.value( QStringLiteral( "name" ) ).toString();
-    QString displayName = serviceName.split( '/' ).last();
-    if ( !baseChecked )
-    {
-      adjustBaseUrl( base, serviceName );
-      baseChecked = true;
-    }
-
-    visitor( displayName, base + serviceName + '/' + serviceType );
-  }
-}
-void QgsArcGisRestUtils::addLayerItems( const std::function< void( const QString &, const QString &, const QString &, const QString &, const QString &, bool, const QString &, const QString & )> &visitor, const QVariantMap &serviceData, const QString &parentUrl, const ServiceTypeFilter filter )
-{
-  const QString authid = QgsArcGisRestUtils::convertSpatialReference( serviceData.value( QStringLiteral( "spatialReference" ) ).toMap() ).authid();
-
-  QString format = QStringLiteral( "jpg" );
-  bool found = false;
-  const QList<QByteArray> supportedFormats = QImageReader::supportedImageFormats();
-  const QStringList supportedImageFormatTypes = serviceData.value( QStringLiteral( "supportedImageFormatTypes" ) ).toString().split( ',' );
-  for ( const QString &encoding : supportedImageFormatTypes )
-  {
-    for ( const QByteArray &fmt : supportedFormats )
-    {
-      if ( encoding.startsWith( fmt, Qt::CaseInsensitive ) )
-      {
-        format = encoding;
-        found = true;
-        break;
-      }
-    }
-    if ( found )
-      break;
-  }
-  const QStringList capabilities = serviceData.value( QStringLiteral( "capabilities" ) ).toString().split( ',' );
-
-  // If the requested layer type is vector, do not show raster-only layers (i.e. non query-able layers)
-  const bool serviceMayHaveQueryCapability = capabilities.contains( QStringLiteral( "Query" ) ) ||
-      serviceData.value( QStringLiteral( "serviceDataType" ) ).toString().startsWith( QLatin1String( "esriImageService" ) );
-  if ( filter == QgsArcGisRestUtils::Vector && !serviceMayHaveQueryCapability )
-    return;
-
-  const QVariantList layerInfoList = serviceData.value( QStringLiteral( "layers" ) ).toList();
-  for ( const QVariant &layerInfo : layerInfoList )
-  {
-    const QVariantMap layerInfoMap = layerInfo.toMap();
-    const QString id = layerInfoMap.value( QStringLiteral( "id" ) ).toString();
-    const QString parentLayerId = layerInfoMap.value( QStringLiteral( "parentLayerId" ) ).toString();
-    const QString name = layerInfoMap.value( QStringLiteral( "name" ) ).toString();
-    const QString description = layerInfoMap.value( QStringLiteral( "description" ) ).toString();
-
-    if ( !layerInfoMap.value( QStringLiteral( "subLayerIds" ) ).toList().empty() )
-    {
-      visitor( parentLayerId, id, name, description, parentUrl + '/' + id, true, QString(), format );
-    }
-    else
-    {
-      visitor( parentLayerId, id, name, description, parentUrl + '/' + id, false, authid, format );
-    }
-  }
-
-  // Add root MapServer as raster layer when multiple layers are listed
-  if ( filter != QgsArcGisRestUtils::Vector && layerInfoList.count() > 1 && serviceData.contains( QStringLiteral( "supportedImageFormatTypes" ) ) )
-  {
-    const QString name = QStringLiteral( "(%1)" ).arg( QObject::tr( "All layers" ) );
-    const QString description = serviceData.value( QStringLiteral( "Comments" ) ).toString();
-    visitor( 0, 0, name, description, parentUrl, false, authid, format );
-  }
-
-  // Add root ImageServer as layer
-  if ( serviceData.value( QStringLiteral( "serviceDataType" ) ).toString().startsWith( QLatin1String( "esriImageService" ) ) )
-  {
-    const QString name = serviceData.value( QStringLiteral( "name" ) ).toString();
-    const QString description = serviceData.value( QStringLiteral( "description" ) ).toString();
-    visitor( 0, 0, name, description, parentUrl, false, authid, format );
-  }
-}
-
-QgsArcGisAsyncQuery::QgsArcGisAsyncQuery( QObject *parent )
-  : QObject( parent )
-{
-}
-
-QgsArcGisAsyncQuery::~QgsArcGisAsyncQuery()
-{
-  if ( mReply )
-    mReply->deleteLater();
-}
-
-void QgsArcGisAsyncQuery::start( const QUrl &url, const QString &authCfg, QByteArray *result, bool allowCache, const QgsStringMap &headers )
-{
-  mResult = result;
-  QNetworkRequest request( url );
-
-  for ( auto it = headers.constBegin(); it != headers.constEnd(); ++it )
-  {
-    request.setRawHeader( it.key().toUtf8(), it.value().toUtf8() );
-  }
-
-  if ( !authCfg.isEmpty() &&  !QgsApplication::authManager()->updateNetworkRequest( request, authCfg ) )
-  {
-    const QString error = tr( "network request update failed for authentication config" );
-    emit failed( QStringLiteral( "Network" ), error );
-    return;
-  }
-
-  QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsArcGisAsyncQuery" ) );
-  if ( allowCache )
-  {
-    request.setAttribute( QNetworkRequest::CacheLoadControlAttribute, QNetworkRequest::PreferCache );
-    request.setAttribute( QNetworkRequest::CacheSaveControlAttribute, true );
-  }
-  mReply = QgsNetworkAccessManager::instance()->get( request );
-  connect( mReply, &QNetworkReply::finished, this, &QgsArcGisAsyncQuery::handleReply );
-}
-
-void QgsArcGisAsyncQuery::handleReply()
-{
-  mReply->deleteLater();
-  // Handle network errors
-  if ( mReply->error() != QNetworkReply::NoError )
-  {
-    QgsDebugMsg( QStringLiteral( "Network error: %1" ).arg( mReply->errorString() ) );
-    emit failed( QStringLiteral( "Network error" ), mReply->errorString() );
-    return;
-  }
-
-  // Handle HTTP redirects
-  QVariant redirect = mReply->attribute( QNetworkRequest::RedirectionTargetAttribute );
-  if ( !redirect.isNull() )
-  {
-    QNetworkRequest request = mReply->request();
-    QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsArcGisAsyncQuery" ) );
-    QgsDebugMsg( "redirecting to " + redirect.toUrl().toString() );
-    request.setUrl( redirect.toUrl() );
-    mReply = QgsNetworkAccessManager::instance()->get( request );
-    connect( mReply, &QNetworkReply::finished, this, &QgsArcGisAsyncQuery::handleReply );
-    return;
-  }
-
-  *mResult = mReply->readAll();
-  mResult = nullptr;
-  emit finished();
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-QgsArcGisAsyncParallelQuery::QgsArcGisAsyncParallelQuery( const QString &authcfg, const QgsStringMap &requestHeaders, QObject *parent )
-  : QObject( parent )
-  , mAuthCfg( authcfg )
-  , mRequestHeaders( requestHeaders )
-{
-}
-
-void QgsArcGisAsyncParallelQuery::start( const QVector<QUrl> &urls, QVector<QByteArray> *results, bool allowCache )
-{
-  Q_ASSERT( results->size() == urls.size() );
-  mResults = results;
-  mPendingRequests = mResults->size();
-  for ( int i = 0, n = urls.size(); i < n; ++i )
-  {
-    QNetworkRequest request( urls[i] );
-    QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsArcGisAsyncParallelQuery" ) );
-    QgsSetRequestInitiatorId( request, QString::number( i ) );
-
-    for ( auto it = mRequestHeaders.constBegin(); it != mRequestHeaders.constEnd(); ++it )
-    {
-      request.setRawHeader( it.key().toUtf8(), it.value().toUtf8() );
-    }
-    if ( !mAuthCfg.isEmpty() && !QgsApplication::authManager()->updateNetworkRequest( request, mAuthCfg ) )
-    {
-      const QString error = tr( "network request update failed for authentication config" );
-      mErrors.append( error );
-      QgsMessageLog::logMessage( error, tr( "Network" ) );
-      continue;
-    }
-
-    request.setAttribute( QNetworkRequest::HttpPipeliningAllowedAttribute, true );
-    if ( allowCache )
-    {
-      request.setAttribute( QNetworkRequest::CacheLoadControlAttribute, QNetworkRequest::PreferCache );
-      request.setAttribute( QNetworkRequest::CacheSaveControlAttribute, true );
-      request.setRawHeader( "Connection", "keep-alive" );
-    }
-    QNetworkReply *reply = QgsNetworkAccessManager::instance()->get( request );
-    reply->setProperty( "idx", i );
-    connect( reply, &QNetworkReply::finished, this, &QgsArcGisAsyncParallelQuery::handleReply );
-  }
-}
-
-void QgsArcGisAsyncParallelQuery::handleReply()
-{
-  QNetworkReply *reply = qobject_cast<QNetworkReply *>( QObject::sender() );
-  QVariant redirect = reply->attribute( QNetworkRequest::RedirectionTargetAttribute );
-  int idx = reply->property( "idx" ).toInt();
-  reply->deleteLater();
-  if ( reply->error() != QNetworkReply::NoError )
-  {
-    // Handle network errors
-    mErrors.append( reply->errorString() );
-    --mPendingRequests;
-  }
-  else if ( !redirect.isNull() )
-  {
-    // Handle HTTP redirects
-    QNetworkRequest request = reply->request();
-    QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsArcGisAsyncParallelQuery" ) );
-    QgsDebugMsg( "redirecting to " + redirect.toUrl().toString() );
-    request.setUrl( redirect.toUrl() );
-    reply = QgsNetworkAccessManager::instance()->get( request );
-    reply->setProperty( "idx", idx );
-    connect( reply, &QNetworkReply::finished, this, &QgsArcGisAsyncParallelQuery::handleReply );
-  }
-  else
-  {
-    // All OK
-    ( *mResults )[idx] = reply->readAll();
-    --mPendingRequests;
-  }
-  if ( mPendingRequests == 0 )
-  {
-    emit finished( mErrors );
-    mResults = nullptr;
-    mErrors.clear();
-  }
-}
-
-///@endcond
